# api/openapi.yaml
openapi: 3.0.0
info:
  title: TravelerTrack API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      required: [code, message]
...
    Location:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
      required: [latitude, longitude]
...
    BuildRouteRequest:
      type: object
      properties:
        tags:
          type: array
          items:
            type: array
            items:
              type: string
          description: Array of tag arrays for each place in the route
        location:
          $ref: '#/components/schemas/Location'
        radius:
          type: number
          format: double
          description: Search radius in meters
          example: 2000
      required: [tags, location, radius]
    GeoJSONFeature:
      type: object
      properties:
        type:
          type: string
          enum: [Feature]
        geometry:
          type: object
          description: GeoJSON Geometry object
        properties:
          type: object
          properties:
            distance:
              type: number
              description: Distance in meters
    RouteSegment:
      type: object
      properties:
        name:
          type: string
          description: Name of the place
        features:
          type: array
          items:
            $ref: '#/components/schemas/GeoJSONFeature'
    RouteBuildResponse:
      type: object
      properties:
        type:
          type: string
          enum: [FeatureCollection]
        segments:
          type: array
          items:
            $ref: '#/components/schemas/RouteSegment'
paths:
...
  /routes/build:
    post:
      summary: Build a route from tags
      operationId: buildRoute
      description: Takes tags, location, and radius to build an optimal route with place recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildRouteRequest'
      responses:
        '200':
          description: Route built successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteBuildResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            text/plain:
              schema:
                type: string
                example: "Sorry, the application-wide daily limit for this feature has been reached."
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'